% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds2aa.R
\name{cds2aa}
\alias{cds2aa}
\title{cds2aa}
\usage{
cds2aa(cds)
}
\arguments{
\item{cds}{\code{DNAStringSet} [mandatory]}
}
\value{
\code{AAStringSet}
}
\description{
This function translates a \code{DNAStringSet} into an \code{AAStringSet}.
}
\examples{
## define two cds sequences
cds1 <- Biostrings::DNAString("ATGCAACATTGC")
cds2 <- Biostrings::DNAString("ATG---CATTGC")
cds1.cds2.aln <- c(Biostrings::DNAStringSet(cds1),
 Biostrings::DNAStringSet(cds2))
cds2aa(cds1.cds2.aln)
}
\seealso{
\code{\link[Biostrings]{XStringSet-class}},
\code{\link[seqinr]{translate}}
}
\author{
Kristian K Ullrich
}
