% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_pairwiseDeletionAA}
\alias{rcpp_pairwiseDeletionAA}
\title{rcpp_pairwiseDeletionAA}
\usage{
rcpp_pairwiseDeletionAA(aavector, ncores = 1L)
}
\arguments{
\item{aavector}{StringVector}

\item{ncores}{number of cores}
}
\description{
returns number of AA sites used
}
\author{
Kristian K Ullrich
}
